New Library
Added New String Library... It is not yet ready to be published in an environment but will be soon

Renamed Initializer_List to BasicInitializer_List to avoid naming conflicts with the compilers.

Had to wrap the String Class in a struct to prevent naming Conflicts. So to use each String you will need Custom_String::String("")